third_party/blink/web_tests/platform/generic/plugins/mouse-click-iframe-to-plugin-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: DidChangeFocus(true)
CONSOLE MESSAGE: Blink Test Plugin: MouseDown at (52,52)
CONSOLE MESSAGE: Blink Test Plugin: MouseUp at (52,52)


PASSED
